Move PresenceCheckAccessor to the accessor package
