+++ processor/src/main/java/org/mapstruct/ap/internal/util/accessor/PresenceCheckAccessor.java
@@ -3,7 +3,7 @@
  *
  * Licensed under the Apache License version 2.0, available at http://www.apache.org/licenses/LICENSE-2.0
  */
-package org.mapstruct.ap.internal.model.common;
+package org.mapstruct.ap.internal.util.accessor;
 
 import javax.lang.model.element.ExecutableElement;
